Predicted Species using Random Forest Classifier Model
color_class Dictionary:

This dictionary maps the species labels of the Iris dataset ('Iris-setosa', 'Iris-versicolor', 'Iris-virginica') to colors that will be used for visualization. For example, 'Iris-setosa' is mapped to 'red', 'Iris-versicolor' to 'green', and 'Iris-virginica' to 'blue'.
Creating a colors List:

The code creates a list called colors by mapping the predicted labels (presumably stored in y_pred_4) to their corresponding colors using the color_class dictionary. This list will be used to color the points in a scatter plot.
Scatter Plot:

The code uses the matplotlib library to create a scatter plot.
It plots the data points from the test dataset (x_test) on a two-dimensional plot where the x-axis represents 'Sepal Length' and the y-axis represents 'Petal Length'.
The c parameter in plt.scatter() specifies the colors for each data point. It uses the colors list created earlier to color each point according to its predicted species.
Axis Labels and Title:

The code sets the labels for the x-axis and y-axis using plt.xlabel() and plt.ylabel().
It also sets the title of the plot using plt.title().
Creating a Legend:

To provide a legend that explains the color-coding, the code creates a legend using plt.legend().
It does so by generating legend entries for each species and their corresponding colors. This is achieved using a list comprehension and plt.Line2D() to create marker entries for each species.
Displaying the Plot:

Finally, plt.show() is used to display the scatter plot with labeled data points and a legend.